pyproject.toml
[tool.poetry]
name = "bot-testing"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
replit = "^2.0.0"
flask = "^1.1.2"
db = "^0.1.1"
discord-pretty-help = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
